Correct some files

+++ Libft/utils/ft_memchr.c
@@ -6,11 +6,13 @@
 /*   By: vcordeir <vcordeir@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/02 21:54:36 by vcordeir          #+#    #+#             */
-/*   Updated: 2021/02/02 21:54:36 by vcordeir         ###   ########.fr       */
+/*   Updated: 2021/02/04 23:35:20 by vcordeir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void    *ft_memchr(const void *s, int c, unsigned int n)
+#include <stddef.h>
+
+void    *ft_memchr(const void *s, int c, size_t n)
 {
     const unsigned char *p = s;
     while (n--)
