+++ Libft/utils/ft_memcpy.c
@@ -6,11 +6,13 @@
 /*   By: vcordeir <vcordeir@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/02 21:54:41 by vcordeir          #+#    #+#             */
-/*   Updated: 2021/02/02 21:54:42 by vcordeir         ###   ########.fr       */
+/*   Updated: 2021/02/04 23:35:47 by vcordeir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void    *ft_memcpy(void *dest, const void *src, unsigned int n)
+#include <stddef.h>
+
+void    *ft_memcpy(void *dest, const void *src, size_t n)
 {
     char *csrc = (char *)src;
     char *cdest = (char *)dest;
